requirements.txt
streamlit
torch
transformers>=4.30
pillow
easyocr
